type transcode handoff manager
